requirements.txt
networkx
numpy
pandas
matplotlib
ortools
peewee
python-dotenv